follow 更新
Fix ball shadow render queue
Reduce size of some networked variables
Improved 8ball rules (pocketing opponents color is NOT ball-in-hand)
Fixed rare case causing ball in hand erroneously (Thx eijis-pan)
Removed sync of cue skin (commented it, add it back yourself if you use it)
Fixed size of Snooker D on 7ft and classic table

+++ Modules/BilliardsModule/UdonScripts/CueController.cs
@@ -17,7 +17,7 @@ public class CueController : UdonSharpBehaviour
     [SerializeField] private GameObject body;
     [SerializeField] private GameObject cuetip;
 
-    [UdonSynced] private int syncedCueSkin;
+    [UdonSynced] private byte syncedCueSkin;
     private int activeCueSkin;
 
     private bool holderIsDesktop;
@@ -319,7 +319,7 @@ public class CueController : UdonSharpBehaviour
         syncedHolderIsDesktop = holderIsDesktop;
         primaryHolding = true;
         primaryLocked = false;
-        syncedCueSkin = table._CanUseCueSkin(Networking.GetOwner(this.gameObject).playerId, syncedCueSkin);
+        syncedCueSkin = (byte)table._CanUseCueSkin(Networking.GetOwner(this.gameObject).playerId, syncedCueSkin);
         //syncedCueSkin = table.activeCueSkin;
         cueScale = cueScaleMine;
         RequestSerialization();
